amended to account for a Mac with no user logged in

+++ Work Scripts (Jamf Pro)/Adobe/Adobe CC/Adobe CC App Upgrade Policy/preinstall.sh
@@ -52,26 +52,22 @@ helperHeading="     Upgrade to ${appNameForInstall}     "
 
 function killAdobe ()
 {
-if [[ "$loggedInUser" == "" ]] || [[ "$loggedInUser" == "root" ]]; then
-    echo "No user logged in"
-else
-    # Get all user Adobe Launch Agents/Daemons PIDs
-    userPIDs=$(su -l "$loggedInUser" -c "/bin/launchctl list | grep adobe" | awk '{print $1}')
-    # Kill all user Adobe Launch Agents/Daemons
-    if [[ "$userPIDs" != "" ]]; then
-        while IFS= read -r line; do
-            kill -9 "$line" 2>/dev/null
-        done <<< "$userPIDs"
-    fi
-    # Unload user Adobe Launch Agents
-    su -l "$loggedInUser" -c "/bin/launchctl unload /Library/LaunchAgents/com.adobe.* 2>/dev/null"
-    # Unload Adobe Launch Daemons
-    /bin/launchctl unload /Library/LaunchDaemons/com.adobe.* 2>/dev/null
-    pkill -9 "obe" >/dev/null 2>&1
-    sleep 5
-    # Close any Adobe Crash Reporter windows (e.g. Bridge)
-    pkill -9 "Crash Reporter" >/dev/null 2>&1
+# Get all user Adobe Launch Agents/Daemons PIDs
+userPIDs=$(su -l "$loggedInUser" -c "/bin/launchctl list | grep adobe" | awk '{print $1}')
+# Kill all user Adobe Launch Agents/Daemons
+if [[ "$userPIDs" != "" ]]; then
+    while IFS= read -r line; do
+        kill -9 "$line" 2>/dev/null
+    done <<< "$userPIDs"
 fi
+# Unload user Adobe Launch Agents
+su -l "$loggedInUser" -c "/bin/launchctl unload /Library/LaunchAgents/com.adobe.* 2>/dev/null"
+# Unload Adobe Launch Daemons
+/bin/launchctl unload /Library/LaunchDaemons/com.adobe.* 2>/dev/null
+pkill -9 "obe" >/dev/null 2>&1
+sleep 5
+# Close any Adobe Crash Reporter windows (e.g. Bridge)
+pkill -9 "Crash Reporter" >/dev/null 2>&1
 }
 
 function jamfHelperConfirm ()
@@ -103,20 +99,8 @@ function jamfHelperDownloadInProgress ()
 ** Download time will depend on the speed of your current internet connection **" -alignDescription natural &
 }
 
-########################################################################
-#                         Script starts here                           #
-########################################################################
-
-# Advise the user what is happening and get confirmation or run anyway in 2 hours time
-jamfHelperConfirm
-# Jamf Helper for app closure and removal
-jamfHelperCleanUp
-# Wait a few seconds for the helper message to be seen before closing the apps
-sleep 5
-# Kill processes to allow uninstall
-killAdobe
-# Wait before uninstalling
-sleep 10
+function removePreviousVersions ()
+{
 echo "Uninstalling previous verisons of ${appNameForRemoval}..."
 # Uninstall 2015 - Look for 2015.1-5 first as they can be uninstalled via command line
 "$binaryPath" --uninstall=1 --sapCode="$sapCode" --baseVersion="$version2015" --platform=osx10-64 --deleteUserPreferences=false >/dev/null 2>&1
@@ -155,9 +139,33 @@ if [[ "$uninstallResult2019" -eq "0" ]]; then
     echo "${appNameForRemoval} 2019 uninstalled"
 fi
 rm -rf "/Applications/${appNameForRemoval} 2019" >/dev/null 2>&1
-# Kill the cleaning up helper
-killall -13 "jamfHelper" >/dev/null 2>&1
-# Jamf Helper for app download+install
-jamfHelperDownloadInProgress
+}
 
+########################################################################
+#                         Script starts here                           #
+########################################################################
+
+if [[ "$loggedInUser" == "" ]] || [[ "$loggedInUser" == "root" ]]; then
+    echo "No user logged in, starting upgrade..."
+    # Remove all previous versions
+    removePreviousVersions
+else
+    echo "Jamf helper displayed to ${loggedInUser} to start the upgrade process"
+    # Advise the user what is happening and get confirmation or run anyway in 2 hours time
+    jamfHelperConfirm
+    # Jamf Helper for app closure and removal
+    jamfHelperCleanUp
+    # Wait a few seconds for the helper message to be seen before closing the apps
+    sleep 5
+    # Kill processes to allow uninstall
+    killAdobe
+    # Wait before uninstalling
+    sleep 10
+    # Remove all previous versions
+    removePreviousVersions
+    # Kill the cleaning up helper
+    killall -13 "jamfHelper" >/dev/null 2>&1
+    # Jamf Helper for app download+install
+    jamfHelperDownloadInProgress
+fi
 exit 0
